requirements.txt
# Core dependencies
requests>=2.31.0
requests-cache>=1.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0

# Face detection and embeddings
insightface>=0.7.3
onnxruntime>=1.16.0

# TMDb API
tmdb-python>=1.0.0

# Data processing
pandas>=2.0.0

# Logging and utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Image deduplication
imagehash>=4.3.1

# Google Images search (optional but recommended)
bing-image-downloader>=1.3.3

# DuckDuckGo images
ddg-python>=1.0.0

# Rate limiting and retry logic
tenacity>=8.2.0

# JSON utilities
jsonschema>=4.19.0

# Testing
pytest>=7.4.0
